Refactor Generic Handling
