+++ kmock-processor/src/test/resources/generatorTest/GenericsSource.kt
@@ -9,9 +9,11 @@ package generatorTest
 import tech.antibytes.kmock.Mock
 
 @Mock(Generics::class)
-interface Generics<K : Any, L> {
+interface Generics<K, L> where L : CharSequence, L : Comparable<L>, K : Any {
     var template: K
     fun <T> foo(payload: T)
-    fun <K: Int> foo(payload: K)
+    fun <T : L> buzz(payload: T)
+    fun <T> bar(payload: T) where T : CharSequence?, T : Map<String, Comparable<List<Array<T>>>>
+    fun <K : Int> foo(payload: K)
 }
